Completed EncryptionKeyManager logic * the key is generated from binary. * encrypted using a JSON table. *Then encrypted again using a JSON table, only now each character is two characters (A = Fe , B = zX)
